requirements.txt
pandas>=2.2.0
streamlit>=1.37.0
python-dotenv>=1.0.1
boto3>=1.34.0
plotly>=5.18.0
google-generativeai>=0.7.0
PyPDF2>=3.0.0
PyMuPDF>=1.23.0
scikit-learn>=1.3.0
numpy>=1.24.0
python-jobspy
joblib>=1.3.0
xgboost>=2.0.0
